--- core/core-api/src/main/kotlin/io/dodn/springboot/CoreApiApplication.kt
@@ -1,13 +1,0 @@
-package io.dodn.springboot
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-
-@ConfigurationPropertiesScan
-@SpringBootApplication
-class CoreApiApplication
-
-fun main(args: Array<String>) {
-    runApplication<CoreApiApplication>(*args)
-}
