--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/controller/v1/ExampleController.kt
@@ -1,35 +1,0 @@
-package io.dodn.springboot.core.api.controller.v1
-
-import io.dodn.springboot.core.api.controller.v1.request.ExampleRequestDto
-import io.dodn.springboot.core.api.controller.v1.response.ExampleResponseDto
-import io.dodn.springboot.core.api.domain.ExampleData
-import io.dodn.springboot.core.api.domain.ExampleService
-import io.dodn.springboot.core.api.support.response.ApiResponse
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class ExampleController(
-    val exampleExampleService: ExampleService
-) {
-    @GetMapping("/get/{exampleValue}")
-    fun exampleGet(
-        @PathVariable exampleValue: String,
-        @RequestParam exampleParam: String
-    ): ApiResponse<ExampleResponseDto> {
-        val result = exampleExampleService.processExample(ExampleData(exampleValue, exampleParam))
-        return ApiResponse.success(ExampleResponseDto(result.data))
-    }
-
-    @PostMapping("/post")
-    fun examplePost(
-        @RequestBody request: ExampleRequestDto
-    ): ApiResponse<ExampleResponseDto> {
-        val result = exampleExampleService.processExample(request.toExampleData())
-        return ApiResponse.success(ExampleResponseDto(result.data))
-    }
-}
