--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/controller/v1/response/ExampleResponseDto.kt
@@ -1,5 +1,0 @@
-package io.dodn.springboot.core.api.controller.v1.response
-
-data class ExampleResponseDto(
-    val result: String
-)
