--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/domain/ExampleData.kt
@@ -1,6 +1,0 @@
-package io.dodn.springboot.core.api.domain
-
-data class ExampleData(
-    val value: String,
-    val param: String
-)
