--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/domain/ExampleResult.kt
@@ -1,5 +1,0 @@
-package io.dodn.springboot.core.api.domain
-
-data class ExampleResult(
-    val data: String
-)
