--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/domain/ExampleService.kt
@@ -1,10 +1,0 @@
-package io.dodn.springboot.core.api.domain
-
-import org.springframework.stereotype.Service
-
-@Service
-class ExampleService() {
-    fun processExample(exampleData: ExampleData): ExampleResult {
-        return ExampleResult(exampleData.value)
-    }
-}
