--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/support/error/CoreApiException.kt
@@ -1,6 +1,0 @@
-package io.dodn.springboot.core.api.support.error
-
-class CoreApiException(
-    val errorType: ErrorType,
-    val data: Any? = null
-) : RuntimeException(errorType.message)
