--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/support/error/ErrorCode.kt
@@ -1,5 +1,0 @@
-package io.dodn.springboot.core.api.support.error
-
-enum class ErrorCode {
-    E500
-}
