--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/support/error/ErrorMessage.kt
@@ -1,13 +1,0 @@
-package io.dodn.springboot.core.api.support.error
-
-data class ErrorMessage private constructor(
-    val code: String,
-    val message: String,
-    val data: Any? = null
-) {
-    constructor(errorType: ErrorType, data: Any? = null) : this(
-        code = errorType.code.name,
-        message = errorType.message,
-        data = data
-    )
-}
