--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/support/error/ErrorType.kt
@@ -1,8 +1,0 @@
-package io.dodn.springboot.core.api.support.error
-
-import org.springframework.boot.logging.LogLevel
-import org.springframework.http.HttpStatus
-
-enum class ErrorType(val status: HttpStatus, val code: ErrorCode, val message: String, val logLevel: LogLevel) {
-    DEFAULT_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, ErrorCode.E500, "An unexpected error has occurred.", LogLevel.ERROR)
-}
