--- core/core-api/src/main/kotlin/io/dodn/springboot/core/api/support/response/ResultType.kt
@@ -1,5 +1,0 @@
-package io.dodn.springboot.core.api.support.response
-
-enum class ResultType {
-    SUCCESS, ERROR
-}
