+++ core/core-api/src/main/kotlin/io/simple/coupon/CoreApiApplication.kt
@@ -1,0 +1,13 @@
+package io.simple.coupon
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.runApplication
+
+@ConfigurationPropertiesScan
+@SpringBootApplication
+class CoreApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<io.simple.coupon.CoreApiApplication>(*args)
+}
