+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/controller/ApiControllerAdvice.kt
@@ -1,0 +1,35 @@
+package io.simple.coupon.core.api.controller
+
+import io.simple.coupon.core.api.support.error.CoreApiException
+import io.simple.coupon.core.api.support.error.ErrorType
+import io.simple.coupon.core.api.support.response.ApiResponse
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import org.springframework.boot.logging.LogLevel
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice
+class ApiControllerAdvice {
+    private val log: Logger = LoggerFactory.getLogger(javaClass)
+
+    @ExceptionHandler(CoreApiException::class)
+    fun handleCoreApiException(e: CoreApiException): ResponseEntity<ApiResponse<Any>> {
+        when (e.errorType.logLevel) {
+            LogLevel.ERROR -> log.error("CoreApiException : {}", e.message, e)
+            LogLevel.WARN -> log.warn("CoreApiException : {}", e.message, e)
+            else -> log.info("CoreApiException : {}", e.message, e)
+        }
+        return ResponseEntity(ApiResponse.error(e.errorType, e.data), e.errorType.status)
+    }
+
+    @ExceptionHandler(Exception::class)
+    fun handleException(e: Exception): ResponseEntity<ApiResponse<Any>> {
+        log.error("Exception : {}", e.message, e)
+        return ResponseEntity(
+            ApiResponse.error(ErrorType.DEFAULT_ERROR),
+            ErrorType.DEFAULT_ERROR.status
+        )
+    }
+}
