+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/controller/v1/request/ExampleRequestDto.kt
@@ -1,0 +1,11 @@
+package io.simple.coupon.core.api.controller.v1.request
+
+import io.simple.coupon.core.api.domain.ExampleData
+
+data class ExampleRequestDto(
+    val data: String
+) {
+    fun toExampleData(): ExampleData {
+        return ExampleData(data, data)
+    }
+}
