+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/controller/v1/response/ExampleResponseDto.kt
@@ -1,0 +1,5 @@
+package io.simple.coupon.core.api.controller.v1.response
+
+data class ExampleResponseDto(
+    val result: String
+)
