+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/domain/ExampleData.kt
@@ -1,0 +1,6 @@
+package io.simple.coupon.core.api.domain
+
+data class ExampleData(
+    val value: String,
+    val param: String
+)
