+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/domain/ExampleResult.kt
@@ -1,0 +1,5 @@
+package io.simple.coupon.core.api.domain
+
+data class ExampleResult(
+    val data: String
+)
