+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/domain/ExampleService.kt
@@ -1,0 +1,10 @@
+package io.simple.coupon.core.api.domain
+
+import org.springframework.stereotype.Service
+
+@Service
+class ExampleService() {
+    fun processExample(exampleData: ExampleData): ExampleResult {
+        return ExampleResult(exampleData.value)
+    }
+}
