+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/support/error/CoreApiException.kt
@@ -1,0 +1,6 @@
+package io.simple.coupon.core.api.support.error
+
+class CoreApiException(
+    val errorType: ErrorType,
+    val data: Any? = null
+) : RuntimeException(errorType.message)
