+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/support/error/ErrorCode.kt
@@ -1,0 +1,5 @@
+package io.simple.coupon.core.api.support.error
+
+enum class ErrorCode {
+    E500
+}
