+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/support/error/ErrorMessage.kt
@@ -1,0 +1,13 @@
+package io.simple.coupon.core.api.support.error
+
+data class ErrorMessage private constructor(
+    val code: String,
+    val message: String,
+    val data: Any? = null
+) {
+    constructor(errorType: ErrorType, data: Any? = null) : this(
+        code = errorType.code.name,
+        message = errorType.message,
+        data = data
+    )
+}
