+++ core/core-api/src/main/kotlin/io/simple/coupon/core/api/support/response/ResultType.kt
@@ -1,0 +1,5 @@
+package io.simple.coupon.core.api.support.response
+
+enum class ResultType {
+    SUCCESS, ERROR
+}
