+++ core/core-api/src/main/resources/application.yml
@@ -2,12 +2,6 @@ spring.application.name: core-api
 spring.profiles.active: local
 
 spring:
-  config:
-    import:
-      - monitoring.yml
-      - logging.yml
-      - db-core.yml
-      - client-example.yml
   mvc.throw-exception-if-no-handler-found: true
   web.resources.add-mappings: false
