--- core/core-api/src/test/kotlin/io/dodn/springboot/CoreApiApplicationTest.kt
@@ -1,10 +1,0 @@
-package io.dodn.springboot
-
-import org.junit.jupiter.api.Test
-
-internal class CoreApiApplicationTest : ContextTest() {
-    @Test
-    fun shouldBeLoadedContext() {
-        // it should be passed
-    }
-}
