--- core/core-api/src/test/kotlin/io/dodn/springboot/DevelopTest.kt
@@ -1,10 +1,0 @@
-package io.dodn.springboot
-
-import org.junit.jupiter.api.Tag
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.TestConstructor
-
-@Tag("develop")
-@SpringBootTest
-@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
-abstract class DevelopTest
