+++ core/core-api/src/test/kotlin/io/simple/coupon/ContextTest.kt
@@ -1,0 +1,10 @@
+package io.simple.coupon
+
+import org.junit.jupiter.api.Tag
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.TestConstructor
+
+@Tag("context")
+@SpringBootTest
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+abstract class ContextTest
