+++ core/core-api/src/test/kotlin/io/simple/coupon/CoreApiApplicationTest.kt
@@ -1,0 +1,10 @@
+package io.simple.coupon
+
+import org.junit.jupiter.api.Test
+
+internal class CoreApiApplicationTest : ContextTest() {
+    @Test
+    fun shouldBeLoadedContext() {
+        // it should be passed
+    }
+}
